eatgo-admin-api/src/test/java/me/spring/restaurant/application/RegionServiceTests.java
package me.spring.restaurant.application;

import me.spring.restaurant.domain.Region;
import me.spring.restaurant.domain.RegionRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.ArrayList;
import java.util.List;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.verify;

class RegionServiceTests {

    private RegionService regionService;

    @Mock
    private RegionRepository regionRepository;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);

        regionService = new RegionService(regionRepository);
    }

    @Test
    public void getRegions() {
        List<Region> mockRegions = new ArrayList<>();
        mockRegions.add(Region.builder().name("Seoul").build());

        given(regionRepository.findAll()).willReturn(mockRegions);

        List<Region> regions = regionService.getRegions();

        Region region = regions.get(0);

        assertThat(region.getName(), is("Seoul"));
    }

    @Test
    public void addRegion() {
        Region region = Region.builder().name("Seoul").build();

        given(regionRepository.save(any())).willReturn(region);

        Region created = regionService.addRegion(region);

        verify(regionRepository).save(any());

        assertThat(created.getName(), is("Seoul"));
    }
}
